+++ crates/web-client/js/types/index.d.ts
@@ -1,4 +1,4 @@
-// import { WebClient as WasmWebClient } from "./crates/miden_client_web";
+import { WebClient as WasmWebClient } from "./crates/miden_client_web";
 
 export {
   Account,
@@ -42,18 +42,27 @@ export {
   TransactionScriptInputPair,
   TransactionScriptInputPairArray,
   Word,
-  WebClient
+  WebClient,
 } from "./crates/miden_client_web";
 
-// // Define WebClient args
-// export type SerializedAccountStorageMode = "private" | "public";
-// export type SerializedNoteType = "private" | "public" | "encrypted";
+// Extend WASM WebClient but override methods that use workers
+export declare class WebClient extends WasmWebClient {
+  /**
+   * Factory method to create and initialize a new wrapped WebClient.
+   *
+   * @param rpcUrl - The RPC URL (optional).
+   * @param proverUrl - The prover URL (optional).
+   * @param seed - The seed for the account (optional).
+   * @returns A promise that resolves to a fully initialized WebClient.
+   */
+  static create_client(
+    rpcUrl?: string,
+    proverUrl?: string,
+    seed?: string
+  ): Promise<WebClient>;
 
-// // Extend WASM WebClient but override methods that use workers
-// export class WebClient extends WasmWebClient {
-//   constructor(...args: any[]);
-//   // new_wallet(storageMode: SerializedAccountStorageMode, mutable: boolean): Promise<Account>;
-//   new_faucet(storageMode: SerializedAccountStorageMode, nonFungible: boolean, tokenSymbol: string, decimals: number, maxSupply: string): Promise<Account>;
-//   // new_mint_transaction(target_account_id: string, faucet_id: string, note_type: SerializedNoteType, amount: string): Promise<string>;
-//   terminate(): void;
-// }
+  /**
+   * Terminates the underlying worker.
+   */
+  terminate(): void;
+}