+++ crates/web-client/js/wasm.js
@@ -1,3 +1,2 @@
-// js/wasm.js
 import * as wasmModule from "../Cargo.toml";
 export default wasmModule;