+++ crates/web-client/src/models/account.rs
@@ -1,9 +1,10 @@
-use miden_objects::account::{Account as NativeAccount, AccountType as NativeAccountType};
 use miden_client::utils::{Deserializable, Serializable};
+use miden_objects::{
+    account::{Account as NativeAccount, AccountType as NativeAccountType},
+    utils::SliceReader,
+};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::js_sys::Uint8Array;
-use miden_crypto::utils::SliceReader;
-use serde::{Serialize, Deserialize};
 
 use super::{
     account_code::AccountCode, account_id::AccountId, account_storage::AccountStorage,
@@ -11,7 +12,6 @@ use super::{
 };
 
 #[wasm_bindgen]
-// #[derive(Serialize, Deserialize)]  // Use Serde for auto serialization
 pub struct Account(NativeAccount);
 
 #[wasm_bindgen]
@@ -61,7 +61,6 @@ impl Account {
     }
 
     pub fn serialize(&self) -> Uint8Array {
-        // Estimate the size for the buffer
         let native_account = &self.0;
         let size_hint = native_account.get_size_hint();
         let mut buffer = vec![0u8; size_hint];
@@ -72,7 +71,8 @@ impl Account {
     pub fn deserialize(bytes: Uint8Array) -> Result<Account, JsValue> {
         let vec: Vec<u8> = bytes.to_vec();
         let mut reader = SliceReader::new(&vec); // Wrap with SliceReader
-        let native_account = NativeAccount::read_from(&mut reader).map_err(|e| JsValue::from_str(&format!("Deserialization error: {:?}", e)));
+        let native_account = NativeAccount::read_from(&mut reader)
+            .map_err(|e| JsValue::from_str(&format!("Deserialization error: {:?}", e)));
         Ok(Account(native_account?))
     }
 }