+++ crates/web-client/src/models/provers.rs
@@ -16,30 +16,36 @@ pub struct TransactionProver {
 impl TransactionProver {
     pub fn new_local_prover() -> TransactionProver {
         let local_prover = LocalTransactionProver::new(Default::default());
-        TransactionProver { prover: Arc::new(local_prover), endpoint: None }
+        TransactionProver {
+            prover: Arc::new(local_prover),
+            endpoint: None,
+        }
     }
 
     pub fn new_remote_prover(endpoint: &str) -> TransactionProver {
         let remote_prover = RemoteTransactionProver::new(endpoint);
-        TransactionProver { prover: Arc::new(remote_prover), endpoint: Some(endpoint.to_string()) }
+        TransactionProver {
+            prover: Arc::new(remote_prover),
+            endpoint: Some(endpoint.to_string()),
+        }
     }
 
-    pub fn from_str(prover_type: &str, endpoint: Option<String>) -> TransactionProver {
+    pub fn serialize(&self) -> String {
+        match &self.endpoint {
+            Some(ep) => format!("remote:{}", ep),
+            None => "local".to_string(),
+        }
+    }
+
+    pub fn deserialize(prover_type: &str, endpoint: Option<String>) -> TransactionProver {
         match prover_type {
             "local" => TransactionProver::new_local_prover(),
             "remote" => {
                 // Use as_deref() to convert Option<String> to Option<&str>
-                let ep = endpoint.as_deref().unwrap_or("http://localhost:8080");
+                let ep = endpoint.as_deref().unwrap_or("http://localhost:50051"); // TODO: is this right?
                 TransactionProver::new_remote_prover(ep)
             },
             _ => panic!("Invalid prover type"), // Consider better error handling in production.
-        }
-    }
-
-    pub fn as_str(&self) -> String {
-        match &self.endpoint {
-            Some(ep) => format!("remote:{}", ep),
-            None => "local".to_string(),
         }
     }