+++ crates/web-client/test/new_account.test.ts
@@ -44,7 +44,8 @@ export const createNewWallet = async (
       }
 
       let client = window.client;
-      const accountStorageMode = window.AccountStorageMode.from_str(_storageMode);
+      const accountStorageMode =
+        window.AccountStorageMode.deserialize(_storageMode);
 
       const newWallet = await client.new_wallet(accountStorageMode, _mutable);
 
@@ -152,7 +153,8 @@ export const createNewFaucet = async (
   return await testingPage.evaluate(
     async (_storageMode, _nonFungible, _tokenSymbol, _decimals, _maxSupply) => {
       const client = window.client;
-      const accountStorageMode = window.AccountStorageMode.from_str(_storageMode);
+      const accountStorageMode =
+        window.AccountStorageMode.deserialize(_storageMode);
       const newFaucet = await client.new_faucet(
         accountStorageMode,
         _nonFungible,